Add EmojiMetaData type
